primitive mutliple objects
