+++ Makefile
@@ -18,7 +18,7 @@ MATH_SRC	= ray vec_util vec_operations vec_double_operations
 MATH		= $(addprefix $(MATH_DIR), $(addsuffix .c, $(MATH_SRC)))
 
 GRAPHICS_DIR	= src/graphics/
-GRAPHICS_SRC	= window draw
+GRAPHICS_SRC	= window draw hit
 GRAPHICS		= $(addprefix $(GRAPHICS_DIR), $(addsuffix .c, $(GRAPHICS_SRC)))
 
 PARSER_DIR		= src/parser/
