+++ src/math/ray.c
@@ -29,13 +29,11 @@ t_vector	at(t_ray ray, double t)
 	return (add_to_vec(ray.origin, dt));
 }
 
-uint32_t	ray_color(t_ray	ray, t_object sp)
+uint32_t	ray_color(t_ray	ray)
 {
 	t_vector	col;
 	double		t;
 
-	if (hit_sphere(sp, ray))
-		return (color(sp.color.x, sp.color.y, sp.color.z, 255.0));
 	t = 0.5 * normalize(ray.direction).y + 1.0;
 	col = add_to_vec(mult_double_vec(1.0 - t, new_vec(255, 255, 255)),
 			mult_double_vec(t, new_vec(127, 200, 255)));
